kvs: changed folder structure
- To enable cargo test again
- Since kvs and kvs_tool have different deps.
