+++ Cargo.toml
@@ -1,7 +1,8 @@
 [workspace]
 resolver = "2"
 members=[
-    "src/rust",
+    "src/rust/rust_kvs",
+    "src/rust/rust_kvs_tool"
 ]
 
 [workspace.package]
@@ -10,7 +11,8 @@ edition = "2021"
 
 
 [workspace.dependencies]
-rust-kvs = { path = "src/rust" }
+rust-kvs = { path = "src/rust/rust_kvs" }
+rust_kvs_tool = { path = "src/rust/rust_kvs_tool" }
 
 adler32 = "1.2.0"
 tinyjson = "2.5.1"
