+++ src/rust/rust_kvs/tests/kvs_check_need_kvs.rs
@@ -9,23 +9,23 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-//! # Verify File Check for non-existing Defaults File
+//! # Verify File Check for non-existing KVS File
 
-use rust_kvs::{ErrorCode, InstanceId, Kvs, OpenNeedDefaults, OpenNeedKvs};
+use rust_kvs::{ErrorCode, InstanceId, Kvs, KvsApi, OpenNeedDefaults, OpenNeedKvs};
 
 mod common;
 use crate::common::TempDir;
 
-/// Start with no KVS and check if the `need_defaults` flag is working
+/// Start with no KVS and check if the `need_kvs` flag is working
 #[test]
-fn kvs_check_needs_defaults() -> Result<(), ErrorCode> {
+fn kvs_check_needs_kvs() -> Result<(), ErrorCode> {
     let dir = TempDir::create()?;
     dir.set_current_dir()?;
 
     let kvs = Kvs::open(
         InstanceId::new(0),
-        OpenNeedDefaults::Required,
-        OpenNeedKvs::Optional,
+        OpenNeedDefaults::Optional,
+        OpenNeedKvs::Required,
     );
 
     assert_eq!(kvs.err(), Some(ErrorCode::KvsFileReadError));
