+++ src/rust/rust_kvs/tests/kvs_snapshot_rotation.rs
@@ -11,7 +11,9 @@
 
 //! # Verify Snapshot Rotation
 
-use rust_kvs::{ErrorCode, InstanceId, Kvs, KvsValue, OpenNeedDefaults, OpenNeedKvs, SnapshotId};
+use rust_kvs::{
+    ErrorCode, InstanceId, Kvs, KvsApi, KvsValue, OpenNeedDefaults, OpenNeedKvs, SnapshotId,
+};
 use std::collections::HashMap;
 
 mod common;
